// packages/common/src/enums.ts
export enum PluralVisibility {
    Trusted = 'Trusted',
    Private = 'Private',
    Public = 'Public',
}

export enum MemberFieldType {
    String = 'String',
    Color = 'Color',
    Date = 'Date',
    Month = 'Month',
    Year = 'Year',
    MonthYear = 'MonthYear',
    Timestamp = 'Timestamp',
    MonthDay = 'MonthDay',
}

export enum UserRole {
    User = 'User',
    Admin = 'Admin',
}